refactor hooks and reducers into modules
