+++ frontend/src/hooks/tasks/index.ts
@@ -1,0 +1,1 @@
+export { useTasks } from "./useTasks";
