+++ frontend/src/reducers/tasks/index.ts
@@ -1,0 +1,1 @@
+export { tasksReducer, initialState, State, Action } from "./tasksReducer";
